+++ BreadPlayer.Views.UWP/Core/SharedLogic.cs
@@ -82,9 +82,9 @@ namespace BreadPlayer.Core
                 if (_player == null)
                 {
                     _player = new BassPlayerEngine(
-                        ApiInformation.IsApiContractPresent("Windows.Phone.PhoneContract", 1), 
+                        ApiInformation.IsApiContractPresent("Windows.Phone.PhoneContract", 1),
                         SettingsHelper.GetRoamingSetting<bool>("CrossfadeEnabled", true),
-                        SettingsHelper.GetRoamingSetting<int>("DeviceBufferSize", 350));
+                        SettingsHelper.GetRoamingSetting<double>("DeviceBufferSize", 350.00));
                 }
 
                 return _player;
